--- src/App.vue
@@ -1,17 +1,0 @@
-<template>
-  <Layout>
-    <router-view />
-  </Layout>
-</template>
-
-<script>
-import { defineComponent } from 'vue'
-import Layout from './components/layout/index.vue'
-
-export default defineComponent({
-  components: {
-    Layout
-  },
-  setup(props) {}
-})
-</script>
