+++ src/app/components/layout/interface.ts
@@ -1,0 +1,3 @@
+import type { BLOG_CONFIG } from '../../../../blog.config'
+
+export type LayoutProvide = BLOG_CONFIG
