+++ src/app/pages/project.vue
