--- src/components/layout/interface.ts
@@ -1,3 +1,0 @@
-import type { BLOG_CONFIG } from "../../../blog.config";
-
-export type LayoutProvide = BLOG_CONFIG;
