+++ src/constants/error.ts
@@ -1,0 +1,3 @@
+export const INVALID_ERROR = 500
+export const NOT_FOUND = 404
+export const FORBIDDEN = 403
