--- src/entry-client.ts
@@ -1,7 +1,0 @@
-import { createApp } from './main'
-
-const { app, router } = createApp()
-
-router.isReady().then(() => {
-  app.mount('#app')
-})
