--- src/entry-server.ts
@@ -1,64 +1,0 @@
-import { createApp } from './main'
-
-import { renderToString } from '@vue/server-renderer'
-import { renderHeadToString } from '@vueuse/head'
-
-import { basename } from 'path'
-
-export const render = async (url: string, manifest) => {
-  const { router, app, head, routes, _document } = createApp()
-  const hasMatch = routes.includes(url)
-  router.push(url)
-  await router.isReady()
-  const ctx = {} as any
-  const html = await renderToString(app, ctx)
-  const document = await renderToString(_document)
-  const preloadLinks = renderPreloadLinks(ctx.modules, manifest)
-  const { headTags } = renderHeadToString(head)
-  return [html, preloadLinks, headTags, hasMatch, document]
-}
-
-function renderPreloadLinks(modules, manifest) {
-  let links = ''
-  const seen = new Set()
-  modules.forEach((id) => {
-    const files = manifest[id]
-    if (files) {
-      files.forEach((file) => {
-        if (!seen.has(file)) {
-          seen.add(file)
-          const filename = basename(file)
-          if (manifest[filename]) {
-            for (const depFile of manifest[filename]) {
-              links += renderPreloadLink(depFile)
-              seen.add(depFile)
-            }
-          }
-          links += renderPreloadLink(file)
-        }
-      })
-    }
-  })
-  return links
-}
-
-function renderPreloadLink(file) {
-  if (file.endsWith('.js')) {
-    return `<link rel="modulepreload" crossorigin href="${file}">`
-  } else if (file.endsWith('.css')) {
-    return `<link rel="stylesheet" href="${file}">`
-  } else if (file.endsWith('.woff')) {
-    return ` <link rel="preload" href="${file}" as="font" type="font/woff" crossorigin>`
-  } else if (file.endsWith('.woff2')) {
-    return ` <link rel="preload" href="${file}" as="font" type="font/woff2" crossorigin>`
-  } else if (file.endsWith('.gif')) {
-    return ` <link rel="preload" href="${file}" as="image" type="image/gif">`
-  } else if (file.endsWith('.jpg') || file.endsWith('.jpeg')) {
-    return ` <link rel="preload" href="${file}" as="image" type="image/jpeg">`
-  } else if (file.endsWith('.png')) {
-    return ` <link rel="preload" href="${file}" as="image" type="image/png">`
-  } else {
-    // TODO
-    return ''
-  }
-}
