+++ src/ssr.ts
@@ -1,0 +1,3 @@
+import { createSSRApp } from 'vue'
+import { createMemoryHistory } from 'vue-router'
+import { renderToString } from '@vue/server-renderer'
