--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './resolver'
-
-export * from './time'
